xorm使用SqlTemplateClient时候有几个坑，当然现在只是模糊的猜测 1. 传入模板的key必须是全名（加上.stpl后缀） 2. 目测只支持Execute, Query方法 3. 如果使用SqlTemplateClient时，使用其它方法（例如InsertOne, Get）则静默使用ORM方式

+++ app/model/article_model.go
@@ -21,7 +21,7 @@ type ArticleMarkRequest struct {
 
 type ArticleMarkResponse struct {
 	BaseResponse
-	MarkTimes int64 `json:"markTimes"`
+	MarkTimes int32 `json:"markTimes"`
 }
 
 type ArticleFavoriteRequest struct {
